rfid status cards

+++ frontend/src/components/activeStatusList/rfidsstatus/FreeRfid.jsx
@@ -1,14 +1,12 @@
 import React, { useEffect, useState } from "react";
-import "./tableStyle.css";
 
-import { createTheme, ThemeProvider } from "@mui/material/styles";
-import Paginator from "../paginator/Paginator";
 import Button from "@mui/material/Button";
 import axios from "axios";
-import MessagePopup from "../messageComponent/MessagePopup";
-import { formatRfidStatus } from "./formatRfidStatus";
+import { formatRfidStatus } from "../../devices/formatRfidStatus";
+import MessagePopup from "../../messageComponent/MessagePopup";
+import Paginator from "../../paginator/Paginator";
 
-const AllRfidList = () => {
+const FreeRfid = () => {
   const [FreeRfidsData, setFreeRfidsData] = useState([]); // State for the full data
   const [currentPage, setCurrentPage] = useState(1); // State for current page
   const [itemsPerPage] = useState(10); // Number of items per page
@@ -52,42 +50,44 @@ const AllRfidList = () => {
   }, []);
 
   // **Filter the deviceData to show only "Active" devices**
-  const allFreeRfids = FreeRfidsData;
+  const inactiveFreeRfids = FreeRfidsData.filter(
+    (rfid) => rfid.rfidStatus === "UNASSIGNED"
+  );
 
   // Calculate the current items to display on the current page
   const indexOfLastItem = currentPage * itemsPerPage;
   const indexOfFirstItem = indexOfLastItem - itemsPerPage;
-  const currentItems = allFreeRfids.slice(indexOfFirstItem, indexOfLastItem); // Slice the filtered data based on current page
+  const currentItems = inactiveFreeRfids.slice(
+    indexOfFirstItem,
+    indexOfLastItem
+  ); // Slice the filtered data based on current page
 
   // Calculate the total number of pages for FreeRfids
-  const totalPages = Math.ceil(allFreeRfids.length / itemsPerPage);
+  const totalPages = Math.ceil(inactiveFreeRfids.length / itemsPerPage);
 
   // Function to handle page change
   const handlePageChange = (pageNumber) => {
     setCurrentPage(pageNumber); // Set the new page number
   };
 
-  const { palette } = createTheme();
-  const { augmentColor } = palette;
-  const createColor = (mainColor) =>
-    augmentColor({ color: { main: mainColor } });
-  const theme = createTheme({
-    palette: {
-      inactivebtn: createColor("#ccc"),
-      assignedColor: createColor("#95EF71"),
-      unAssignedColor: createColor("#F6B85A"),
-    },
-  });
-
-  const handelRfidDelete = async (id) => {
+  const handelRfidDelete = async (id, status) => {
     try {
-      await axios.delete(
-        `${process.env.REACT_APP_API_URL}/api/delete/rfidkey/${id}`
-      );
+      // Check if the RFID status is not 'Unassigned'
+      if (status !== "Unassigned") {
+        // If the status is not 'Unassigned', proceed with the deletion
+        await axios.delete(
+          `${process.env.REACT_APP_API_URL}/api/delete/rfidkey/${id}`
+        );
 
-      addMessage("Rfid Deleted Successfully!", "success");
-      // Remove the deleted RFID from the local state
-      setFreeRfidsData((prevData) => prevData.filter((rfid) => rfid.id !== id));
+        addMessage("Rfid Deleted Successfully!", "success");
+
+        // Remove the deleted RFID from the local state
+        setFreeRfidsData((prevData) =>
+          prevData.filter((rfid) => rfid.id !== id)
+        );
+      } else {
+        addMessage("Cannot delete RFID with status 'Unassigned'", "error");
+      }
     } catch (err) {
       if (err.response) {
         const errorMessage =
@@ -104,15 +104,14 @@ const AllRfidList = () => {
   return (
     <div>
       <h2 className="tableDataHeaderTitle">
-        <span></span> All Rfid Keys
+        <span>{inactiveFreeRfids.length}</span> Unassigned Rfid Keys
       </h2>
       <table border="1" cellPadding="10" className="activedevicesTable">
         <thead className="activedevicesTable-header">
           <tr>
             <th>RFID Keys</th>
-            <th>Status</th>
             <th>Tag Type</th>
-            <th>Installation Date</th>
+            <th>Status</th>
             <th>Delete</th>
           </tr>
         </thead>
@@ -121,57 +120,28 @@ const AllRfidList = () => {
             ? currentItems.map((Rfids) => (
                 <tr key={Rfids.id}>
                   <td>{Rfids.keyCode}</td>
-                  <td>
-                    <ThemeProvider theme={theme}>
-                      <Button
-                        variant="contained"
-                        color={
-                          Rfids.rfidStatus === "ASSIGNED"
-                            ? "assignedColor"
-                            : "unAssignedColor"
-                        }
-                        className="smallbutton"
-                      >
-                        <span className="sentencebutton">
-                          {formatRfidStatus(Rfids.rfidStatus)}
-                        </span>
-                      </Button>
-                    </ThemeProvider>
-                  </td>
                   <td>{formatRfidStatus(Rfids.rfidType)}</td>
+                  <td>{formatRfidStatus(Rfids.rfidStatus)}</td>
                   <td>
-                    {Rfids.installationDate
-                      ? new Date(Rfids.installationDate)
-                          .toISOString()
-                          .split("T")[0]
-                      : "Not installed"}
-                  </td>
-
-                  <td>
-                    <ThemeProvider theme={theme}>
-                      <Button
-                        variant="contained"
-                        color={
-                          Rfids.rfidStatus === "ASSIGNED"
-                            ? "inactivebtn"
-                            : "error"
+                    <Button
+                      variant="contained"
+                      color="error"
+                      className="smallbutton"
+                    >
+                      <span
+                        className="sentencebutton"
+                        onClick={() =>
+                          handelRfidDelete(Rfids.id, Rfids.rfidStatus)
                         }
-                        className="smallbutton"
-                        onClick={() => {
-                          if (Rfids.rfidStatus === "ASSIGNED") {
-                            addMessage("The Rfid Is Assigned", "warning");
-                          } else {
-                            handelRfidDelete(Rfids.id);
-                          }
-                        }}
                       >
-                        <span className="sentencebutton">Delete</span>
-                      </Button>
-                    </ThemeProvider>
+                        Delete
+                      </span>
+                    </Button>
                   </td>
                 </tr>
               ))
             : null}
+          {/* No need to display "No active devices available" */}
         </tbody>
       </table>
 
@@ -187,4 +157,4 @@ const AllRfidList = () => {
   );
 };
 
-export default AllRfidList;
+export default FreeRfid;
